fixed img2img and negative prompts
